# specs/002-admin-data-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Admin Data Management API
  version: 1.0.0
servers:
  - url: /api/admin
security:
  - AdminToken: []
paths:
  /skills:
    get:
      summary: List skills
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: "#/components/schemas/Skill"
    post:
      summary: Create skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /skills/{skillId}:
    get:
      summary: Get skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404":
          description: Not found
    put:
      summary: Update skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
        - name: ifMatch
          in: header
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "409":
          description: Conflict (stale update)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      summary: Soft delete skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "409":
          description: Conflict (skill in use)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
  /skills/{skillId}/restore:
    post:
      summary: Restore skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /scenarios:
    get:
      summary: List scenarios
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarios:
                    type: array
                    items:
                      $ref: "#/components/schemas/Scenario"
    post:
      summary: Create scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /scenarios/{scenarioId}:
    get:
      summary: Get scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
    put:
      summary: Update scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
        - name: ifMatch
          in: header
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "409":
          description: Conflict (stale update)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
    delete:
      summary: Soft delete scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "409":
          description: Conflict (scenario has sessions)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
  /scenarios/{scenarioId}/restore:
    post:
      summary: Restore scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /scenarios/{scenarioId}/publish:
    post:
      summary: Publish scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /scenarios/{scenarioId}/unpublish:
    post:
      summary: Unpublish scenario
      parameters:
        - name: scenarioId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /sessions:
    get:
      summary: List sessions
      parameters:
        - name: scenarioId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionSummary"
  /sessions/{sessionId}:
    get:
      summary: Get session detail
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "404":
          description: Not found
    delete:
      summary: Delete session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /audit-log:
    get:
      summary: List audit log entries
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: adminId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  schemas:
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        rubric:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, category, rubric, status, createdAt, updatedAt]
    SkillInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        rubric:
          type: string
        description:
          type: string
      required: [name, category, rubric]
    Scenario:
      type: object
      properties:
        id:
          type: string
        metadata:
          type: object
          properties:
            title:
              type: string
            slug:
              type: string
            domain:
              type: string
            scenarioType:
              type: string
            difficulty:
              type: string
            conflictLevel:
              type: string
            estimatedDurationMinutes:
              type: integer
            tags:
              type: array
              items:
                type: string
        context:
          type: object
          properties:
            situation:
              type: string
            background:
              type: string
            setting:
              type: string
        simulationConfig:
          type: object
          properties:
            ai:
              $ref: "#/components/schemas/Persona"
            trainee:
              $ref: "#/components/schemas/Persona"
            language:
              type: string
            conversationStart:
              type: object
              properties:
                speakerRoleId:
                  type: string
                initialPromptToUser:
                  type: string
            conversationRules:
              type: object
              properties:
                stayInCharacter:
                  type: boolean
                allowNarration:
                  type: boolean
                coachingAllowed:
                  type: boolean
                tone:
                  type: string
            conversationDynamics:
              type: object
              properties:
                typicalBehaviors:
                  type: array
                  items:
                    type: string
                possibleResponses:
                  type: array
                  items:
                    type: string
            decisionConstraints:
              type: object
              additionalProperties: true
            conversationEndConditions:
              type: object
              properties:
                possibleEndStates:
                  type: array
                  items:
                    type: string
        evaluationConfig:
          type: object
          properties:
            learningObjectives:
              type: array
              items:
                type: string
            evaluationCriteria:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  description:
                    type: string
            skillsAssessed:
              type: array
              items:
                type: string
            scoring:
              type: object
              additionalProperties: true
            evaluationInstructionsForLLM:
              type: string
        status:
          type: string
          enum: [draft, published]
        recordStatus:
          type: string
          enum: [active, deleted]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ScenarioInput:
      type: object
      properties:
        metadata:
          type: object
          properties:
            title:
              type: string
            slug:
              type: string
            domain:
              type: string
            scenarioType:
              type: string
            difficulty:
              type: string
            conflictLevel:
              type: string
            estimatedDurationMinutes:
              type: integer
            tags:
              type: array
              items:
                type: string
        context:
          type: object
          properties:
            situation:
              type: string
            background:
              type: string
            setting:
              type: string
        simulationConfig:
          type: object
          properties:
            ai:
              $ref: "#/components/schemas/Persona"
            trainee:
              $ref: "#/components/schemas/Persona"
            language:
              type: string
            conversationStart:
              type: object
              properties:
                speakerRoleId:
                  type: string
                initialPromptToUser:
                  type: string
            conversationRules:
              type: object
              properties:
                stayInCharacter:
                  type: boolean
                allowNarration:
                  type: boolean
                coachingAllowed:
                  type: boolean
                tone:
                  type: string
            conversationDynamics:
              type: object
              properties:
                typicalBehaviors:
                  type: array
                  items:
                    type: string
                possibleResponses:
                  type: array
                  items:
                    type: string
            decisionConstraints:
              type: object
              additionalProperties: true
            conversationEndConditions:
              type: object
              properties:
                possibleEndStates:
                  type: array
                  items:
                    type: string
        evaluationConfig:
          type: object
          properties:
            learningObjectives:
              type: array
              items:
                type: string
            evaluationCriteria:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  description:
                    type: string
            skillsAssessed:
              type: array
              items:
                type: string
            scoring:
              type: object
              additionalProperties: true
            evaluationInstructionsForLLM:
              type: string
        status:
          type: string
          enum: [draft, published]
    Persona:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        personality:
          type: array
          items:
            type: string
        motivations:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        tendencies:
          type: array
          items:
            type: string
        knowledge:
          type: array
          items:
            type: string
        emotionalState:
          type: string

    SessionSummary:
      type: object
      properties:
        id:
          type: string
        scenarioId:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        terminationReason:
          type: string
          nullable: true
        evaluationStatus:
          type: string
      required: [id, scenarioId, status, startedAt, evaluationStatus]
    SessionDetail:
      type: object
      properties:
        id:
          type: string
        scenarioId:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        terminationReason:
          type: string
          nullable: true
        evaluationStatus:
          type: string
        transcript:
          type: array
          items:
            $ref: "#/components/schemas/Turn"
        evaluation:
          $ref: "#/components/schemas/Evaluation"
      required: [id, scenarioId, status, startedAt, evaluationStatus, transcript]
    Turn:
      type: object
      properties:
        sequence:
          type: integer
        speaker:
          type: string
        transcript:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [sequence, speaker, createdAt]
    Evaluation:
      type: object
      properties:
        status:
          type: string
        scores:
          type: array
          items:
            $ref: "#/components/schemas/Score"
        summary:
          type: string
          nullable: true
      required: [status]
    Score:
      type: object
      properties:
        skillId:
          type: string
        rating:
          type: integer
        note:
          type: string
      required: [skillId, rating, note]
    AuditEntry:
      type: object
      properties:
        id:
          type: string
        adminId:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: string
          nullable: true
      required: [id, adminId, action, entityType, entityId, timestamp]
    ValidationError:
      type: object
      properties:
        error:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      required: [error]
    ConflictError:
      type: object
      properties:
        error:
          type: string
        impactedIds:
          type: array
          items:
            type: string
      required: [error, impactedIds]